laag components refactor
