+++ app/(authenticated)/user/feed/page.tsx
@@ -8,7 +8,7 @@ import { useAvatar } from "@/hooks/useAvatar"
 import { useLaagImage } from "@/hooks/useLaagImage"
 import Image from "next/image"
 import { format } from "date-fns"
-import { EditLaagDialog } from "../groups/[id]/edit-laag-dialog"
+import { EditLaagDialog } from "../../../../components/laags/edit-laag-dialog"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
