+++ app/(authenticated)/user/groups/[id]/page.tsx
@@ -10,10 +10,10 @@ import { Users, Edit2, Settings, User2 } from "lucide-react"
 import { useGroupPicture } from "@/hooks/useGroupPicture"
 import { useAvatar } from "@/hooks/useAvatar"
 import Image from "next/image"
-import { CreateLaagDialog } from "./create-laag-dialog"
-import { LaagFeed } from "./laag-feed"
-import { EditGroupModal } from "@/components/groups/edit-group-modal"
-import { ManageMembersDialog } from "./manage-members-dialog"
+import { CreateLaagDialog } from "../../../../../components/laags/create-laag-dialog"
+import { LaagFeed } from "../../../../../components/laags/laag-feed"
+import { EditGroupModal } from "@/components/groups/edit-group-dialog"
+import { ManageMembersDialog } from "../../../../../components/laags/manage-members-dialog"
 import { Button } from "@/components/ui/button"
 import { useAuth } from "@/app/context/auth-context"
 import { toast } from "sonner"
