+++ components/groups/group-table.tsx
@@ -21,8 +21,8 @@ import { useAvatar } from "@/hooks/useAvatar"
 import Image from "next/image"
 import { toast } from "sonner"
 import { useAuth } from "@/app/context/auth-context"
-import { NewGroupDialog } from "@/components/groups/new-group-dialog"
-import { EditGroupModal } from "@/components/groups/edit-group-modal"
+import { NewGroupDialog } from "@/components/groups/create-group-dialog"
+import { EditGroupModal } from "@/components/groups/edit-group-dialog"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { useRouter } from "next/navigation"
